Add usage right supplier configuration
