+++ common-lib/src/main/resources/application.conf
@@ -93,5 +93,7 @@ usageRightsConfigProvider = {
       "CC BY-SA-4.0",
       "CC BY-ND-4.0"
     ]
+    freeSuppliers = []
+    suppliersCollectionExcl {}
   }
 }
