+++ common-lib/src/main/scala/com/gu/mediaservice/lib/guardian/GuardianUsageRightsConfig.scala
@@ -128,4 +128,88 @@ object GuardianUsageRightsConfig extends UsageRightsConfigProvider {
     "CC BY-4.0", "CC BY-SA-4.0", "CC BY-ND-4.0"
   )
 
+  /* These are currently hardcoded */
+  val payGettySourceList = List(
+    "Arnold Newman Collection",
+    "360cities.net Editorial",
+    "360cities.net RM",
+    "age fotostock RM",
+    "Alinari",
+    "Arnold Newman Collection",
+    "ASAblanca",
+    "Bob Thomas Sports Photography",
+    "Carnegie Museum of Art",
+    "Catwalking",
+    "Contour",
+    "Contour RA",
+    "Corbis Premium Historical",
+    "Editorial Specials",
+    "Reportage Archive",
+    "Gamma-Legends",
+    "Genuine Japan Editorial Stills",
+    "Genuine Japan Creative Stills",
+    "George Steinmetz",
+    "Getty Images Sport Classic",
+    "Iconic Images",
+    "Iconica",
+    "Icon Sport",
+    "Kyodo News Stills",
+    "Lichfield Studios Limited",
+    "Lonely Planet Images",
+    "Lonely Planet RF",
+    "Masters",
+    "Major League Baseball Platinum",
+    "Moment Select",
+    "Mondadori Portfolio Premium",
+    "National Geographic",
+    "National Geographic RF",
+    "National Geographic Creative",
+    "National Geographic Magazines",
+    "NBA Classic",
+    "Neil Leifer Collection",
+    "Newspix",
+    "PA Images",
+    "Papixs",
+    "Paris Match Archive",
+    "Paris Match Collection",
+    "Pele 10",
+    "Photonica",
+    "Photonica World",
+    "Popperfoto",
+    "Popperfoto Creative",
+    "Premium Archive",
+    "Reportage Archive",
+    "SAMURAI JAPAN",
+    "Sports Illustrated",
+    "Sports Illustrated Classic",
+    "Sygma Premium",
+    "Terry O'Neill",
+    "The Asahi Shimbun Premium",
+    "The LIFE Premium Collection",
+    "ullstein bild Premium",
+    "Ulrich Baumgarten",
+    "VII Premium",
+    "Vision Media",
+    "Xinhua News Agency"
+  )
+
+  val freeSuppliers = List(
+    "AAP",
+    "Alamy",
+    "Allstar Picture Library",
+    "AP",
+    "EPA",
+    "Getty Images",
+    "PA",
+    "Reuters",
+    "Rex Features",
+    "Ronald Grant Archive",
+    "Action Images",
+    "Action Images/Reuters"
+  )
+
+  val suppliersCollectionExcl = Map(
+    "Getty Images" -> payGettySourceList
+  )
+
 }
