+++ common-lib/src/test/resources/application.conf
@@ -52,5 +52,7 @@ usageRightsConfigProvider = {
     contractIllustrators = []
     staffIllustrators = []
     creativeCommonsLicense = []
+    freeSuppliers = []
+    suppliersCollectionExcl {}
   }
 }
